GKE R599. Alpha release version 20100830. Polar not implemented. Tricopter scaling on yaw removed. 20Hz yaw gyro filtering.

+++ telemetry.c
@@ -100,7 +100,7 @@ void SendFlightPacket(void)
 	TxESCi8((int8)AltComp);
 
 	for ( b = 0; b < 6; b++ ) // motor/servo channels
-	 	TxESCu8(PWM[b]);
+	 	TxESCu8((uint8)PWM[b]);
 
 	TxESCi24(mSClock());
 } //
